Auto-sync Plaid data on TRANSACTIONS webhooks

+++ api/plaid/item.js
@@ -1,8 +1,10 @@
 import { handleDeletePlaidItem } from "../../server/plaid/handlers.js";
+import { enforceRateLimit, generalApiRateLimit } from "../../server/http/rateLimit.js";
 import { applySecurityHeaders, assertMethod } from "../../server/http/responseHelpers.js";
 
-export default function handler(request, response) {
+export default async function handler(request, response) {
   applySecurityHeaders(response);
   if (!assertMethod(request, response, "DELETE")) return;
+  if (!(await enforceRateLimit(request, response, generalApiRateLimit))) return;
   return handleDeletePlaidItem(request, response);
 }
